+++ server/src/main.rs
@@ -136,7 +136,11 @@ impl StreamHandler<Result<ws::Message, ws::ProtocolError>> for GameSocket {
                             }
                             "/use" => {
                                 let t = chunks[2].parse::<i32>().unwrap();
-                                let target = if t == -1 { None } else { Some(t) };
+                                let target = if t == -1 {
+                                    None
+                                } else {
+                                    Some(t as usize)
+                                };
                                 use_item(idx, target, &mut self.ecs);
                             }
                             _ => ()
